requirements.txt
discord.py
async_cleverbot

